remove SAS tokens (#99)
Removing the AZURE_URL_EXPIRATION_SECS stops an expiry being propogated to the azurestorage class. This prevents the SAS token from being created for the resources.

This change happened because the tokens are causing images to load incorrectly. This change will be reverted when we create a fix for this issue with SAS tokens